refacto: déplacement de fonctions dédiées à l'Index EgaPro dans un fichier spécifique
