+++ impact/reglementations/models/__init__.py
@@ -1,5 +1,6 @@
 from .bdse import *  # noqa
 from .csrd import *  # noqa
+from .index_egapro import *  # noqa
 
 # This uncommon structure of modules (included in a heigh-level package)
 # can be useful for domains / apps with a big number of models and/or complex ones.
