+++ impact/reglementations/models/bdse.py
@@ -11,20 +11,6 @@ from entreprises.models import Entreprise
 from utils.models import TimestampedModel
 
 
-def derniere_annee_a_publier_index_egapro():
-    annee = datetime.date.today().year
-    return annee - 1
-
-
-def prochaine_echeance_index_egapro(derniere_annee_est_publiee: bool) -> datetime.date:
-    aujourdhui = datetime.date.today()
-    annee = aujourdhui.year
-    if derniere_annee_est_publiee:
-        return datetime.date(annee + 1, 3, 1)
-    else:
-        return datetime.date(annee, 3, 1)
-
-
 def derniere_annee_a_remplir_bdese():
     annee = datetime.date.today().year
     return annee - 1
